+++ test/ctor.test.js
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2012, Trent Mick. All rights reserved.
+ *
+ * Test type checking on creation of the Logger.
+ */
+
+var test = require('tap').test;
+var Logger = require('../lib/bunyan');
+
+
+test("douglas adams", function(t) {
+  t.ok(42, "the answer to life the universe and everything");
+  t.end();
+});
+
+